Improvements but not good balance

+++ lib/pid/pid.h
@@ -9,7 +9,7 @@ struct pid
     double m_integral_sum;
     double m_last_error;
     double m_desired_value;
-    uint32_t m_previous_time;
+    int64_t m_previous_time;
     double m_max_value;
     double m_min_value;
     uint8_t m_stop_windup;
@@ -20,9 +20,9 @@ struct pid pid_init(
     double gain_integral, 
     double gain_derivative, 
     double desired_value,
-    uint32_t time,
+    int64_t time,
     double max_value,
     double min_value,
     uint8_t stop_windup);
-double pid_get_error(struct pid* pid_instance, double value, uint32_t time);
+double pid_get_error(struct pid* pid_instance, double value, int64_t time);
 void pid_set_desired_value(struct pid* pid_instance, double value);
